requirements.txt
# Digital Game Marketplace - Python Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0

# Database
sqlalchemy>=2.0.0

# Data Validation
pydantic>=2.5.0
email-validator>=2.1.0  # Required for Pydantic EmailStr

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# Form Data Handling
python-multipart>=0.0.6

# Development (optional)
# pytest>=8.0.0
# httpx>=0.26.0

